LEONI_LAD_TRACKER_BACK/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
PyMySQL==1.1.0
SQLAlchemy~=2.0.31
typing_extensions==4.11.0
Werkzeug~=3.0.3
zipp==3.18.1

pyodbc~=5.1.0
requests~=2.32.3